Moved common configs to root folder. Fixed some issues. You can now eslint the workspace (manually)
