+++ packages/ptolemy-arktype-modeling/eslint.config.mjs
@@ -1,2 +1,2 @@
-import commonConfig from '../../common-config/eslint.config.mjs';
+import { commonConfig } from '../../eslint.config.mjs';
 export default commonConfig(import.meta.dirname);
