+++ packages/ptolemy-web/.prettierrc.mjs
@@ -1,1 +1,1 @@
-export { default as default } from '../../common-config/prettierrc.mjs';
+export { default as default } from '../../prettierrc.mjs';
