+++ tsconfig.json
@@ -1,0 +1,58 @@
+{
+    "include": [
+        "**/*.*"
+    ],
+    "exclude": [
+        "packages/**/*",
+        "example/**/*",
+        "tools/**/*"
+    ],
+    "compilerOptions": {
+        "importHelpers": true,
+        "target": "ESNext",
+        "useDefineForClassFields": true,
+        "lib": [
+            "ES2022",
+            "DOM",
+            "DOM.Iterable", 
+            "ES2021.WeakRef"
+        ],
+        "module": "esnext",
+        /* Bundler mode */
+        "moduleResolution": "bundler",
+        "resolveJsonModule": true,
+        "isolatedModules": true,
+        "noEmit": true,
+        "declaration": true,
+        /*"jsx": "react-jsx",*/
+        "incremental": true,
+        "esModuleInterop": true,
+        "forceConsistentCasingInFileNames": true /* Ensure that casing is correct in imports. */,
+        /* Type Checking */
+        "strict": true /* Enable all strict type-checking options. */,
+        "noImplicitAny": true /* Enable error reporting for expressions and declarations with an implied 'any' type. */,
+        "strictNullChecks": true /* When type checking, take into account 'null' and 'undefined'. */,
+        "strictFunctionTypes": true /* When assigning functions, check to ensure parameters and the return values are subtype-compatible. */,
+        "strictBindCallApply": true /* Check that the arguments for 'bind', 'call', and 'apply' methods match the original function. */,
+        "strictPropertyInitialization": true /* Check for class properties that are declared but not set in the constructor. */,
+        "noImplicitThis": true /* Enable error reporting when 'this' is given the type 'any'. */,
+        "useUnknownInCatchVariables": true /* Default catch clause variables as 'unknown' instead of 'any'. */,
+        "alwaysStrict": true /* Ensure 'use strict' is always emitted. */,
+        "exactOptionalPropertyTypes": true /* Interpret optional property types as written, rather than adding 'undefined'. */,
+        "noImplicitReturns": true /* Enable error reporting for codepaths that do not explicitly return in a function. */,
+        "noFallthroughCasesInSwitch": true /* Enable error reporting for fallthrough cases in switch statements. */,
+        "noUncheckedIndexedAccess": true /* Add 'undefined' to a type when accessed using an index. */,
+        "noImplicitOverride": true /* Ensure overriding members in derived classes are marked with an override modifier. */,
+        "noPropertyAccessFromIndexSignature": true /* Enforces using indexed accessors for keys declared using an indexed type. */,
+        // "allowUnusedLabels": true,                        /* Disable error reporting for unused labels. */
+        // "allowUnreachableCode": true,                     /* Disable error reporting for unreachable code. */
+
+        /* Completeness */
+        "skipLibCheck": true, /* Skip type checking all .d.ts files. */
+        
+        "types": ["vitest/globals", "vite/client"],
+
+        "allowJs": true,
+        "outDir": "dist" // NOT USED
+    }
+}
